refactor: extract cross-encoder rerank into _RerankMixin
Third decomposition cut. Moves the Phase D sub-D-1 reranker path —
_compute_rerank_state / _maybe_rerank_compact / _flatten_archive_hits /
_redistribute_reranked_hits / _maybe_rerank_search_all plus the
_RERANKER_* event-name constants and _INFO_LEVEL_TELEMETRY_EVENTS — into
openzim_mcp/rerank.py as _RerankMixin.

The telemetry event-name constants are shared between the rerank methods
AND the core _track / handle_zim_query snapshot logic, so they live in
rerank.py and are imported back into simple_tools (the dependency already
runs simple_tools → rerank, so no circular import; the reverse would have
created one). External deps are just _track / zim_operations / _telemetry,
declared under TYPE_CHECKING.

Pure code movement, no behavior change. simple_tools.py 5,358 → 5,194
(5,896 → 5,194 across the three mixin cuts; -702 lines / -12%). Full suite
green (2623 passed); mypy/flake8/isort clean.
